requirements.txt
python-telegram-bot[job-queue]==21.9
SQLAlchemy==2.0.36
aiosqlite==0.19.0
python-dotenv==1.0.0
aiohttp==3.9.1
cryptography==42.0.0
greenlet==3.1.1
cryptography
cryptography
